chore: refactor naming of channel feature
