+++ lib/app/features/chat/community/channel/providers/channel_provider.c.dart
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: ice License 1.0
 
 import 'package:ion/app/exceptions/exceptions.dart';
+import 'package:ion/app/features/chat/community/data/channel_type.dart';
 import 'package:ion/app/features/chat/community/data/entities/community_definition_data.c.dart';
 import 'package:ion/app/features/chat/model/channel_admin_type.dart';
-import 'package:ion/app/features/chat/model/channel_type.dart';
 import 'package:ion/app/features/chat/providers/channel_admins_provider.c.dart';
 import 'package:ion/app/features/ion_connect/model/event_setting.c.dart';
 import 'package:ion/app/features/ion_connect/model/file_alt.dart';
@@ -99,7 +99,7 @@ class ChannelNotifier extends _$ChannelNotifier {
       if (patchChannelResult == null || editChannelResult == null) {
         throw FailedToEditChannelException();
       }
-      return null;
+      return;
     });
   }
