+++ lib/app/features/chat/community/channel/views/create_channel_modal/components/channel_photo.dart
