+++ lib/app/features/chat/community/channel/views/create_channel_modal/create_channel_modal.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/extensions/extensions.dart';
 import 'package:ion/app/features/chat/community/channel/views/components/channel_form.dart';
-import 'package:ion/app/features/chat/views/pages/new_channel_modal/components/channel_photo.dart';
+import 'package:ion/app/features/chat/community/channel/views/create_channel_modal/components/channel_photo.dart';
 import 'package:ion/app/router/app_routes.c.dart';
 import 'package:ion/app/router/components/navigation_app_bar/navigation_app_bar.dart';
 import 'package:ion/app/router/components/navigation_app_bar/navigation_close_button.dart';
