+++ lib/app/features/chat/community/data/channel_type.dart
