+++ lib/app/features/chat/community/view/pages/admins_management_modal/admins_management_modal.dart
@@ -7,9 +7,9 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/components/button/button.dart';
 import 'package:ion/app/components/screen_offset/screen_side_offset.dart';
 import 'package:ion/app/extensions/extensions.dart';
+import 'package:ion/app/features/chat/community/view/pages/add_admin_modal/add_admin_modal.dart';
+import 'package:ion/app/features/chat/community/view/pages/admins_management_modal/components/admin_card.dart';
 import 'package:ion/app/features/chat/providers/channel_admins_provider.c.dart';
-import 'package:ion/app/features/chat/views/pages/new_channel_modal/pages/add_admin_modal/add_admin_modal.dart';
-import 'package:ion/app/features/chat/views/pages/new_channel_modal/pages/admins_management_modal/components/admin_card.dart';
 import 'package:ion/app/router/components/navigation_app_bar/navigation_app_bar.dart';
 import 'package:ion/app/router/components/navigation_app_bar/navigation_close_button.dart';
 import 'package:ion/app/router/utils/show_simple_bottom_sheet.dart';
