+++ lib/app/features/chat/model/channel_data.c.dart
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: ice License 1.0
 
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:ion/app/features/chat/community/data/channel_type.dart';
 import 'package:ion/app/features/chat/model/channel_admin_type.dart';
-import 'package:ion/app/features/chat/model/channel_type.dart';
 import 'package:ion/app/services/media_service/media_service.c.dart';
 
 part 'channel_data.c.freezed.dart';
