+++ lib/app/features/ion_connect/model/tags/community_admin_tag.c.dart
@@ -11,7 +11,7 @@ class CommunityAdminTag with _$CommunityAdminTag {
   const CommunityAdminTag._();
 
   factory CommunityAdminTag.fromTags(List<List<String>> tags) {
-    final tag = tags.where((tag) => tag[0] == tagName && tag[2] == roleValue);
+    final tag = tags.where((tag) => tag[0] == tagName && tag[3] == roleValue);
 
     if (tag.isEmpty) {
       return const CommunityAdminTag(values: []);
@@ -24,6 +24,6 @@ class CommunityAdminTag with _$CommunityAdminTag {
   static const String roleValue = 'admin';
 
   List<List<String>> toTag() {
-    return values.map((value) => [tagName, value, roleValue]).toList();
+    return values.map((value) => [tagName, value, '', roleValue]).toList();
   }
 }
